requirements.txt
django
djangorestframework
azure-monitor-opentelemetry
python-dotenv